requirements.txt

streamlit==1.40.0
pandas>=2.2.3
requests==2.31.0
scipy>=1.16.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.11
google-generativeai>=0.8.0
python-dotenv==1.0.1
ddgs>=0.1.0
beautifulsoup4==4.12.3
bcrypt==4.1.3
lxml==5.2.2